+++ application/src/main/java/com/hashmapinc/server/controller/AssetController.java
@@ -36,7 +36,7 @@ import com.hashmapinc.server.common.data.asset.AssetSearchQuery;
 import com.hashmapinc.server.common.data.security.Authority;
 import com.hashmapinc.server.dao.exception.IncorrectParameterException;
 import com.hashmapinc.server.dao.model.ModelConstants;
-import com.hashmapinc.server.exception.TempusException;
+import com.hashmapinc.server.common.data.exception.TempusException;
 
 import java.util.ArrayList;
 import java.util.List;
