+++ application/src/main/java/com/hashmapinc/server/controller/BaseController.java
@@ -59,9 +59,9 @@ import com.hashmapinc.server.dao.rule.RuleService;
 import com.hashmapinc.server.dao.user.UserService;
 import com.hashmapinc.server.dao.widget.WidgetTypeService;
 import com.hashmapinc.server.dao.widget.WidgetsBundleService;
-import com.hashmapinc.server.exception.TempusErrorCode;
+import com.hashmapinc.server.common.data.exception.TempusErrorCode;
 import com.hashmapinc.server.exception.TempusErrorResponseHandler;
-import com.hashmapinc.server.exception.TempusException;
+import com.hashmapinc.server.common.data.exception.TempusException;
 import com.hashmapinc.server.service.component.ComponentDiscoveryService;
 import com.hashmapinc.server.service.kubernetes.gateway.TempusGatewayKubernetesService;
 import com.hashmapinc.server.service.metadataingestion.MetadataConfigService;
