+++ application/src/main/java/com/hashmapinc/server/controller/ComponentDescriptorController.java
@@ -18,7 +18,7 @@ package com.hashmapinc.server.controller;
 
 import com.hashmapinc.server.common.data.plugin.ComponentDescriptor;
 import com.hashmapinc.server.common.data.plugin.ComponentType;
-import com.hashmapinc.server.exception.TempusException;
+import com.hashmapinc.server.common.data.exception.TempusException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
