+++ application/src/main/java/com/hashmapinc/server/controller/DeviceController.java
@@ -29,8 +29,8 @@ import com.hashmapinc.server.dao.depthseries.DepthSeriesService;
 import com.hashmapinc.server.dao.exception.IncorrectParameterException;
 import com.hashmapinc.server.dao.model.ModelConstants;
 import com.hashmapinc.server.dao.timeseries.TimeseriesService;
-import com.hashmapinc.server.exception.TempusErrorCode;
-import com.hashmapinc.server.exception.TempusException;
+import com.hashmapinc.server.common.data.exception.TempusErrorCode;
+import com.hashmapinc.server.common.data.exception.TempusException;
 import com.hashmapinc.server.service.security.model.SecurityUser;
 import com.opencsv.CSVWriter;
 import lombok.extern.slf4j.Slf4j;
