+++ application/src/main/java/com/hashmapinc/server/controller/EntityRelationController.java
@@ -22,8 +22,8 @@ import com.hashmapinc.server.common.data.relation.EntityRelation;
 import com.hashmapinc.server.common.data.relation.EntityRelationInfo;
 import com.hashmapinc.server.common.data.relation.EntityRelationsQuery;
 import com.hashmapinc.server.common.data.relation.RelationTypeGroup;
-import com.hashmapinc.server.exception.TempusErrorCode;
-import com.hashmapinc.server.exception.TempusException;
+import com.hashmapinc.server.common.data.exception.TempusErrorCode;
+import com.hashmapinc.server.common.data.exception.TempusException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
