+++ application/src/main/java/com/hashmapinc/server/controller/MetadataController.java
@@ -32,8 +32,8 @@ import com.hashmapinc.server.common.data.metadata.MetadataQuery;
 import com.hashmapinc.server.common.data.metadata.MetadataQueryId;
 import com.hashmapinc.server.common.data.page.TextPageData;
 import com.hashmapinc.server.common.data.page.TextPageLink;
-import com.hashmapinc.server.exception.TempusErrorCode;
-import com.hashmapinc.server.exception.TempusException;
+import com.hashmapinc.server.common.data.exception.TempusErrorCode;
+import com.hashmapinc.server.common.data.exception.TempusException;
 import com.hashmapinc.server.requests.IngestMetadataRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.joda.time.DateTime;
