+++ application/src/main/java/com/hashmapinc/server/controller/TempusGatewayConfigurationController.java
@@ -20,7 +20,7 @@ import com.hashmapinc.server.common.data.TempusGatewayConfiguration;
 import com.hashmapinc.server.common.data.id.TempusGatewayConfigurationId;
 import com.hashmapinc.server.common.data.id.TenantId;
 import com.hashmapinc.server.common.data.kubernetes.ReplicaSetStatus;
-import com.hashmapinc.server.exception.TempusException;
+import com.hashmapinc.server.common.data.exception.TempusException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
