+++ application/src/main/java/com/hashmapinc/server/controller/UserSettingsController.java
@@ -19,13 +19,16 @@ package com.hashmapinc.server.controller;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.hashmapinc.server.common.data.*;
+import com.hashmapinc.server.common.data.Logo;
+import com.hashmapinc.server.common.data.Theme;
+import com.hashmapinc.server.common.data.UserSettings;
 import com.hashmapinc.server.common.data.id.UserId;
 import com.hashmapinc.server.common.data.security.Authority;
 import com.hashmapinc.server.dao.logo.LogoService;
 import com.hashmapinc.server.dao.settings.UserSettingsService;
 import com.hashmapinc.server.dao.theme.ThemeService;
-import com.hashmapinc.server.exception.TempusException;
-import com.hashmapinc.server.service.mail.MailService;
+import com.hashmapinc.server.common.data.exception.TempusException;
+import com.hashmapinc.server.dao.mail.MailService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
